Close LGVISIUM-60: Added elevation extraction
